-Fixed warnings
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@4177 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ CodeCompletionsTests/CCTest/tester.cpp
@@ -47,8 +47,8 @@ void Tester::RunTests()
 	if(success == totalTests) {
 		printf("    All tests passed successfully!!\n");
 	} else {
-		printf("    %u of %u tests passed\n", success, totalTests);
-		printf("    %u of %u tests failed\n", errors, totalTests);
+		printf("    %u of %u tests passed\n", (int)success, (int)totalTests);
+		printf("    %u of %u tests failed\n", (int)errors,  (int)totalTests);
 	}
 }
